requirements.txt
blinker==1.8.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Hashing==1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.3
reportlab==4.2.5
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.3
